+++ docs/src/modules/components/ApiPage/common/classes.ts
@@ -1,0 +1,16 @@
+import kebabCase from 'lodash/kebabCase';
+
+type HashParams = { componentName: string; className: string };
+
+export function getClassesHash({ componentName, className }: HashParams) {
+  return `${kebabCase(componentName)}-classes-${className}`;
+}
+
+export interface ClassDefinition {
+  className: string;
+  key: string;
+  description?: string;
+  isGlobal?: boolean;
+  isDeprecated?: boolean;
+  deprecationInfo?: string;
+}
